Split code in separate files
