+++ index.html
@@ -47,7 +47,7 @@
 
 
   <!-- JS Scripts -->
-  <script src="assets/Script/main.js"></script>
+  <script type="module" src="assets/Script/main.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW" crossorigin="anonymous">
     </script>
